move each component to its own file
